Updated for Python 3
 - Changed to use new Tkinter (tkinter). [Breaks py2k support]
 - Changed raw_input to input. [Breaks py2k support]
 - Removed LiveWires name from module except to provide appropriate
   attribution.
 - Fixed raise syntax (as per PEP 3109).
 - Changed instances of British "Colour" to "Color"
 - Updated setup.py to move pygame from a required to optional
   dependency.
 - Updated __init__.py to use a relative link to the beginners
   submodule. [Relevant: PEP 404]

Colour to Color rationale:
A lot of code libraries and programing languages themselves (think css)
use the US International standard of spelling which defines "Color" is
the correct spelling. Getting students to get used to a british spelling
is counter-productive and will cause issues when programming in most
other languages where "color" is a language construct or "color" is a
library's property.

Pygame Optional Dependency rationale:
At this current stage, PyPi when using http://pygame.org/downloads.shtml
or http://pygame.org/ftp/ to download pypi packages resolves the
incorrect package (1.6.2 instead of 1.9.1). As a result this would cause
breaking python 3 compatibility of this library.

Furthermore as the beginners submodule doesn't require pygame, this
library can be effectively used for basic game making without the use of
pygame.